qwen2_5_omni_encoder_change_BACKUP/spatpy/pyproject.toml
[project]
name = "spatpy"
description = "Spatial audio utilities"
version = "2022.1046"
authors = [
    { name = "Capture Research" }
]
requires-python = ">=3.7"
dependencies = [
    "ninja",
    "numpy>=1.19",
    "plotly>=5.1",
    "xarray>=0.16",
    "scipy>=1.5",
    "numpy-stl",
    "dataclasses",
    "stl-to-voxel==0.9.2"
]

[project.scripts]
autoalign = "spatpy.dsp:autoalign_main"
autoeq = "spatpy.eq:main"
damfeezle = "spatpy.io:damfeezle"
mixmate = "spatpy.io:mixmate"
rosellify = "spatpy.binaural:rosellify"
sample_uniform = "spatpy.uniform_sampling:main"
wav2pv = "spatpy.signal_path.ufb_backend:wav2pv"

[project.optional-dependencies]
docs = [
    "devpi-client",
    "bumpver",
    "sphinx>=4.2",
    "sphinx-plotly-directive>=0.1.3",
    "gitpython",
    "sphinxcontrib-mermaid>=0.7",
    "sphinxcontrib-katex",
    "sphinx-rtd-theme",
    "sphinx_autodoc_typehints",
    "sphinx-argparse",
]
ufb_banding = ["ufb_banding>=2023.0"]
extras = [
    "flit>=3.7",
    "pydub>=0.25",
    "h5netcdf>=0.13",
    "pytest",
    "sympy>=1.9",
    "torch"
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
